Move everything from home into a different component, style fixes&changes
