+++ src/components/ui/bento-grid.tsx
@@ -32,6 +32,7 @@ const BentoCard = ({
 	description,
 	href,
 	cta,
+	transitions = true,
 }: {
 	name: string;
 	className: string;
@@ -40,29 +41,33 @@ const BentoCard = ({
 	description: string;
 	href: string;
 	cta?: string;
+	transitions?: boolean;
 }) => (
 	<div
 		key={name}
 		className={cn(
-			"group relative flex flex-col justify-between overflow-hidden rounded-xl transition-shadow duration-300 ease-in-out",
-			// light styles
-			"bg-white [box-shadow:0_0_0_1px_rgba(0,0,0,.03),0_2px_4px_rgba(0,0,0,.05),0_12px_24px_rgba(0,0,0,.05)]",
-			// dark styles
-			"transform-gpu dark:bg-[var(--color-background)] dark:[border:1px_solid_rgba(255,255,255,.1)] dark:hover:[box-shadow:0_-20px_20px_-20px_#ffffff1f_inset]",
+			"group relative flex flex-col justify-between overflow-hidden rounded-xl transition-shadow duration-300 ease-in-out bg-white dark:bg-[var(--color-background)] border-[1px] border-neutral-200 dark:[border:1px_solid_rgba(255,255,255,.1)] ",
+			transitions && "transform-gpu [box-shadow:0_0_0_1px_rgba(0,0,0,.03),0_2px_4px_rgba(0,0,0,.05),0_12px_24px_rgba(0,0,0,.05)] dark:hover:[box-shadow:0_-20px_20px_-20px_#ffffff1f_inset]",
 			className,
 		)}
 	>
 		{background && <div>{background}</div>}
 		<div className={cn("pointer-events-none z-10 flex transform-gpu flex-col gap-1 p-6 transition-all duration-300",
-			cta
-				? "group-hover:-translate-y-10"
-				: "group-hover:scale-90"
+			cta && transitions && "group-hover:-translate-y-10"
+
 		)}>
-			{<icon.element className={cn("h-12 w-12 origin-left transform-gpu text-neutral-700 transition-all duration-300 ease-in-out group-hover:scale-75", icon.className ? icon.className : "")} {...icon.props} />}
+			{<icon.element
+			className={cn(
+				"h-12 w-12 origin-left text-neutral-700",
+				transitions && "group-hover:scale-75 transition-all duration-300 ease-in-out transform-gpu",
+				icon.className ? icon.className : ""
+			)}
+			{...icon.props}
+			/>}
 			<h3 className="text-xl font-semibold text-neutral-700 dark:text-neutral-300">
 				{name}
 			</h3>
-			<p className="max-w-lg text-neutral-400">{description}</p>
+			<p className="max-w-lg text-neutral-600 dark:text-neutral-400">{description}</p>
 		</div>
 
 		{cta &&
